requirements.txt
torch>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0
wfdb>=4.0.0
pycryptodome>=3.17
numpy>=1.23.0
matplotlib>=3.7.0